treewide: switch to MySQL
This deprecates sqlite. This is a *breaking change* that requires an
existing MySQL database to be set up with all the data migrated.

Thankfully the existing schema isn't very complex so you can dump the
databse to an SQL file and import it into MySQL without much trouble.

The previous dbupdate should already have handled whatever changes are
needed to make this transition smooth.